+++ StarforgeInitiative/Ships/Breadnaught/Cards/BreadnaughtBasicOverdriveCard.cs
@@ -28,9 +28,9 @@ internal sealed class BreadnaughtBasicOverdriveCard : Card, IRegisterable
 	public override CardData GetData(State state)
 		=> upgrade switch
 		{
-			Upgrade.A => new() { cost = 0, artTint = "ff3366" },
-			Upgrade.B => new() { cost = 1, exhaust = true, artTint = "ff3366" },
-			_ => new() { cost = 1, artTint = "ff3366" },
+			Upgrade.B => new() { cost = 1, temporary = true, exhaust = true, artTint = "ff3366" },
+			Upgrade.A => new() { cost = 0, temporary = true, artTint = "ff3366" },
+			_ => new() { cost = 1, temporary = true, artTint = "ff3366" },
 		};
 
 	public override List<CardAction> GetActions(State s, Combat c)