+++ StarforgeInitiative/Ships/Breadnaught/Cards/BreadnaughtBasicPushCard.cs
@@ -28,8 +28,8 @@ internal sealed class BreadnaughtBasicPushCard : CannonColorless, IRegisterable
 	public override CardData GetData(State state)
 		=> upgrade switch
 		{
-			Upgrade.A => new() { cost = 2, flippable = true, artTint = "ff3366" },
-			_ => new() { cost = 2, artTint = "ff3366" },
+			Upgrade.A => new() { cost = 2, temporary = true, flippable = true, artTint = "ff3366" },
+			_ => new() { cost = 2, temporary = true, artTint = "ff3366" },
 		};
 
 	public override List<CardAction> GetActions(State s, Combat c)